+++ src/requires/apclass.lua
@@ -42,6 +42,7 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
     local OrbitTargetSet = false
     local OrbitTargetPlanet = nil
     local OrbitTicks = 0
+    local apRoute = {}
 
     function ap.GetAutopilotBrakeDistanceAndTime(speed)
         return GetAutopilotBrakeDistanceAndTime(speed)
@@ -1010,6 +1011,12 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
                     --horizontalDistance = msqrt(targetVec:len()^2-(coreAltitude-targetAltitude)^2)
                 end
                 if (CustomTarget and CustomTarget.planetname == "Space" and velMag < 50) then
+                    if #apRoute>0 then
+                        BrakeIsOn = false
+                        AP.ToggleAutopilot()
+                        AP.ToggleAutopilot()
+                        return
+                    end
                     finishAutopilot("Autopilot complete, arrived at space location")
                     BrakeIsOn = true
                     brakeInput = 1
@@ -1457,6 +1464,11 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
                     if (not spaceLaunch and not Reentry and distanceToTarget <= brakeDistance and -- + (velMag*deltaTick)/2
                             (constructVelocity:project_on_plane(worldVertical):normalize():dot(targetVec:project_on_plane(worldVertical):normalize()) > 0.99  or VectorStatus == "Finalizing Approach")) then 
                         VectorStatus = "Finalizing Approach" 
+                        if #apRoute>0 then
+                            AP.ToggleAutopilot()
+                            AP.ToggleAutopilot()
+                            return
+                        end
                         AP.cmdThrottle(0) -- Kill throttle in case they weren't in cruise
                         if AltitudeHold then
                             -- if not OrbitAchieved then
@@ -1801,7 +1813,7 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
             end
             VectorStatus = "Proceeding to Waypoint"
         end
-
+        local routeOrbit = false
         if (time - apDoubleClick) < 1.5 and atmosDensity > 0 then
             if not SpaceEngines then
                 msgText = "No space engines detected, Orbital Hop not supported"
@@ -1822,8 +1834,19 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
         end
         TargetSet = false -- No matter what
         -- Toggle Autopilot, as long as the target isn't None
-        if AutopilotTargetIndex > 0 and not Autopilot and not VectorToTarget and not spaceLaunch and not IntoOrbit then
+        if (AutopilotTargetIndex > 0 or #apRoute>0) and not Autopilot and not VectorToTarget and not spaceLaunch and not IntoOrbit then
             if 0.5 * Nav:maxForceForward() / core.g() < coreMass then  msgText = "WARNING: Heavy Loads may affect autopilot performance." msgTimer=5 end
+            if #apRoute>0 and not finalLand then 
+                AutopilotTargetIndex = apRoute[1]
+                ATLAS.UpdateAutopilotTarget()
+                table.remove(apRoute,1)
+                msgText = "Route Autopilot in Progress"
+                local targetVec = CustomTarget.position - worldPos
+                local distanceToTarget = targetVec:project_on_plane(worldVertical):len()
+                if distanceToTarget > 50000 and CustomTarget.planetname == planet.name then 
+                    routeOrbit=true
+                end
+            end
             ATLAS.UpdateAutopilotTarget() -- Make sure we're updated
             AP.showWayPoint(autopilotTargetPlanet, AutopilotTargetCoords)
 
@@ -1844,6 +1867,9 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
                         if not VectorToTarget then
                             play("vtt", "AP")
                             ToggleVectorToTarget(SpaceTarget)
+                            if routeOrbit then
+                                HoldAltitude = planet.noAtmosphericDensityAltitude + LowOrbitHeight
+                            end
                         end
                     else
                         play("apOn", "AP")
@@ -1901,6 +1927,18 @@ function APClass(Nav, core, unit, system, atlas, vBooster, hover, telemeter_1, a
             play("apOff", "AP")
             AP.ResetAutopilots(1)
         end
+    end
+
+    function ap.routeWP(getRoute, clear)
+        if getRoute then return apRoute end
+        if clear then 
+            apRoute = {}
+        else
+            apRoute[#apRoute+1]=AutopilotTargetIndex
+            msgText = "Added "..CustomTarget.name.." to route. "
+            p("Added "..CustomTarget.name.." to route. Total Route: "..json.encode(apRoute))
+        end
+        return apRoute
     end
 
     function ap.cmdThrottle(value, dontSwitch) -- sets the throttle value to value, also switches to throttle mode (vice cruise) unless dontSwitch passed