+++ src/requires/controlclass.lua
@@ -315,6 +315,14 @@ function ControlClass(Nav, core, unit, system, atlas, vBooster, hover, antigrav,
             end
         elseif action == "option8" then
             toggleView = false
+            if AltIsOn and holdingShift then 
+                if AutopilotTargetIndex > 0 and CustomTarget ~= nil then
+                    AP.routeWP()
+                else
+                    msgText = "Select a saved wp on IPH to add to or remove from route"
+                end
+                return
+            end
             stablized = not stablized
             if not stablized then
                 msgText = "DeCoupled Mode - Ground Stabilization off"